backend-app/face/requirements.txt
fastapi
uvicorn
numpy
face_recognition
python-multipart
pillow
